# perf(matrices): Use DomainMatrix.inv in Matrix.inv