--- a/sympy/polys/matrices/rref.py
+++ b/sympy/polys/matrices/rref.py
@@ -207,8 +207,9 @@ def _dm_rref_GJ_sparse(M):
 
 def _dm_rref_GJ_dense(M):
     """Compute RREF using dense Gauss-Jordan elimination with division."""
+    partial_pivot = M.domain.is_RR or M.domain.is_CC
     ddm = M.rep.to_ddm().copy()
-    pivots = ddm_irref(ddm)
+    pivots = ddm_irref(ddm, _partial_pivot=partial_pivot)
     M_rref_ddm = DDM(ddm, M.shape, M.domain)
     pivots = tuple(pivots)
     return M.from_rep(M_rref_ddm.to_dfm_or_ddm()), pivots
@@ -251,6 +252,10 @@ def _dm_rref_choose_method(M, method, *, denominator=False):
             method = _dm_rref_choose_method_ZZ(M, denominator=denominator)
         elif K.is_QQ:
             method = _dm_rref_choose_method_QQ(M, denominator=denominator)
+        elif K.is_RR or K.is_CC:
+            # TODO: Add partial pivot support to the sparse implementations.
+            method = 'GJ'
+            use_fmt = 'dense'
         elif K.is_EX and M.rep.fmt == 'dense' and not denominator:
             # Do not switch to the sparse implementation for EX because the
             # domain does not have proper canonicalization and the sparse