<!-- Your title above should be a short description of what
was changed. Do not include the issue number in the title. -->

#### References to other Issues or PRs
<!-- If this pull request fixes an issue, write "Fixes #NNNN" in that exact
format, e.g. "Fixes #1234" (see
https://tinyurl.com/auto-closing for more information). Also, please
write a comment on that issue linking back to this pull request once it is
open. -->

See the discussion in gh-25448.

This is built on top of gh-25443.

#### Brief description of what is fixed or changed

This just adds one commit on top of gh-25443 (fd675887ec) to enable using `DomainMatrix.inv()` by default in `Matrix.inv()`. Two `method=` options are added to `Matrix.inv`: `DM` for use DomainMatrix and `DMNC` for use DomainMatrix without cancellation. The `DM` option is used by default if a non-EX domain is found for the expressions in the matrix. Otherwise `inv` falls back on the current `GE` method.

Further work would be needed to determine exactly what is the best DomainMatrix algorithm to use for any given matrix. This would be like the analysis in gh-25410: essentially the same issues apply for sparse vs dense and small/big denominators but there are more parameters to consider. A new consideration though is whether or not to cancel the denominator into the matrix because for some domains this is extremely expensive. Another consideration is that we can use `adj_det` instead of `inv_den` which is fully division free and is probably fastest for very small matrices with very complicated entries. Without doing that kind of analysis though I expect that this will make almost all symbolic inverse operations a lot faster than the default method on master and I believe that it should almost always give the same output in the same form.

It is difficult to give timings comparisons because the cases where this helps are essentially ones where the current `Matrix.inv` method (with `GE`) takes basically forever (I don't want to wait long enough to measure how long it takes).

The matrix from gh-25403 can be inverted in parts with this if `method='DMNC'` is used:
```python
In [3]: for p in M2parts:
   ...:     %time ok = p.inv('DMNC')
   ...: 
CPU times: user 1.33 s, sys: 0 ns, total: 1.33 s
Wall time: 1.33 s
CPU times: user 326 ms, sys: 0 ns, total: 326 ms
Wall time: 326 ms
CPU times: user 5.48 ms, sys: 0 ns, total: 5.48 ms
Wall time: 5.49 ms
CPU times: user 307 ms, sys: 0 ns, total: 307 ms
Wall time: 307 ms
CPU times: user 4.27 ms, sys: 0 ns, total: 4.27 ms
Wall time: 4.28 ms
```
With master the 3rd and 5th matrices can be inverted in 200ms (rather than 5ms with the PR). The other 3 matrices take at least 1 minute with master but possibly hours or days or worse - I didn't wait to find out.

Note that those two matrices that can be inverted with master are 1x1 matrices:
```python
In [3]: for p in M2parts:
   ...:     print(p.shape)
   ...: 
(5, 5)
(4, 4)
(1, 1)
(4, 4)
(1, 1)
```
That's right: it takes 200ms to invert a 1x1 matrix! (This is basically the time taken to determine if the single entry in the matrix is zero or not). 

The example in gh-19920 completes in 50ms with the PR:
```python
In [4]: from sympy import symbols, Matrix
   ...: 
   ...: a, b, c, d = symbols('a b c d')
   ...: 
   ...: Z = Matrix([
   ...:     [1/b + 1/a,       -1/b,         0],
   ...:     [     -1/b,  1/c + 1/b,      -1/c],
   ...:     [        0,       -1/c, 1/d + 1/c]])

In [5]: %time Z.inv()
CPU times: user 51.1 ms, sys: 3.98 ms, total: 55.1 ms
Wall time: 54.9 ms
Out[5]: 
⎡a⋅b + a⋅c + a⋅d        a⋅c + a⋅d              a⋅d      ⎤
⎢───────────────      ─────────────       ───────────── ⎥
⎢ a + b + c + d       a + b + c + d       a + b + c + d ⎥
⎢                                                       ⎥
⎢   a⋅c + a⋅d     a⋅c + a⋅d + b⋅c + b⋅d     a⋅d + b⋅d   ⎥
⎢ ─────────────   ─────────────────────   ───────────── ⎥
⎢ a + b + c + d       a + b + c + d       a + b + c + d ⎥
⎢                                                       ⎥
⎢      a⋅d              a⋅d + b⋅d        a⋅d + b⋅d + c⋅d⎥
⎢ ─────────────       ─────────────      ───────────────⎥
⎣ a + b + c + d       a + b + c + d       a + b + c + d ⎦
```
I don't know how long that takes with master (the issue claims more than an hour).

#### Other comments

More work would be needed to make this as good as possible by selecting the best algorithms (rref vs rref_den vs clearing denominators and whether to attempt cancellation by default for different domains etc). I expect though that this is already a massive improvement over master in almost all cases and I would like to consider just merging this change as it is and making further improvements later.

The key questions are:

- Are there are any cases where this causes things to slow down or changes the output in a bad way and is there any possibility of being a bit more conservative about whether to use `DomainMatrix` or not to avoid causing any problems in those cases?
- What should the user interface be? Here I added `DM` and `DMNC` method options to `Matrix.inv`. Users will need at least to be able to distinguish those two cases (to cancel or not) but I am open to any suggestions for what the API should be to do that.

One thing that I do want to add is using the connected components decomposition in `DomainMatrix.inv()`: then it would be possible to e.g. invert the matrix from gh-25403 fully (with `method='DMNC'`). Also that would speed up sparse matrices enormously and could detect many non-invertible matrices without doing any arithmetic (e.g. `M1parts[0]` from gh-25403).

#### Release Notes

<!-- Write the release notes for this release below between the BEGIN and END
statements. The basic format is a bulleted list with the name of the subpackage
and the release note for this PR. For example:

* solvers
  * Added a new solver for logarithmic equations.

* functions
  * Fixed a bug with log of integers. Formerly, `log(-x)` incorrectly gave `-log(x)`.

* physics.units
  * Corrected a semantical error in the conversion between volt and statvolt which
    reported the volt as being larger than the statvolt.

or if no release note(s) should be included use:

NO ENTRY

See https://github.com/sympy/sympy/wiki/Writing-Release-Notes for more
information on how to write release notes. The bot will check your release
notes automatically to see if they are formatted correctly. -->

<!-- BEGIN RELEASE NOTES -->
* matrices
    * The `Matrix.inv()` method now uses DomainMatrix by default when a suitable domain can be found for the elements of a matrix. This makes computing matrix inverses enormously faster for common cases like a matrix with polynomial or rational function entries or Gaussian integers etc. Many matrix inverses that would previously take effectively forever to compute can now be computed in a reasonable amount of time.
<!-- END RELEASE NOTES -->