@sylee957 2023-07-31
I'm a bit concerned about the cases like `(x + 1)**1000` which isn't very easy to check whether even if that is slow, because domain would expand that eagerly. Do you think that this performs better, or that has other solutions, or we can ignore such cases?