@oscarbenjamin 2023-07-31
> I'm a bit concerned about the cases like `(x + 1)**1000`

The first thing to recognise is that this is something that already happens with the current `inv` method:
```python
In [1]: M = Matrix([[(x + 1)**100000]])

In [2]: M.inv(method='GE')
^C^C
KeyboardInterrupt
```
You can try that with all 5 methods and it happens because regardless of the algorithm used by each method the first step is always to check invertibility:
https://github.com/sympy/sympy/blob/e8ffcf40a6148235667ed75b811bdc1f953466e9/sympy/matrices/inverse.py#L187-L195

So from a speed perspective the change here would not make that case *worse* compared to the existing behaviour on master. In fact it performs the expansion a lot faster so if anything it reduces this problem:
```python
In [2]: M = Matrix([[(x + 1)**100]])

In [3]: %time ok = M.inv(method='DM')
CPU times: user 137 ms, sys: 0 ns, total: 137 ms
Wall time: 137 ms

In [4]: %time ok = M.inv(method='GE')
CPU times: user 625 ms, sys: 7.4 ms, total: 633 ms
Wall time: 632 ms
```
The other aspect is the form of output returned and the change here preserves existing behaviour at least for the default `inv` method:
```python
In [6]: M.inv(method='GE')
Out[6]: 
⎡         1         ⎤
⎢───────────────────⎥
⎢ 3      2          ⎥
⎣x  + 3⋅x  + 3⋅x + 1⎦

In [7]: M.inv(method='DM')
Out[7]: 
⎡         1         ⎤
⎢───────────────────⎥
⎢ 3      2          ⎥
⎣x  + 3⋅x  + 3⋅x + 1⎦
```
So comparing the PR to master the expansion problem is reduced from a performance perspective and also the PR preserves the existing form of the output.

In general yes of course we need some limit to expansion but this is really a very general problem because there are a lot of functions that will do this e.g. `as_poly`, `cancel`, `expand`, etc. Unless you are very careful it is difficult to work with something like `(x+1)**100000` because so many functions might expand it. Probably the best thing would be to have e.g. `construct_domain` recognise where it can replace something like that with a new symbol rather than expanding but that adds a lot of complexity in many places and cannot handle something like
```python
In [9]: M = Matrix([[(x + 1)**10000, (x-1)**10000]])
```
Here if we don't expand we will not be able to tell if some combination of these expressions is equal to zero.