@asmeurer 2023-07-31
One instance that's currently extremely slow is the inverse of the general symbolic n x n matrix. The size of this grows like the O(n!) but it should still be representable for n < 10 or so. But right now, n=3 is slow (!), and it can't even compute n = 4. 

Consider:

```
MatrixSymbol("A", 4, 4).as_explicit().inv()
```

This is still slow in this branch (as in hangs, I don't know how long it would take to complete) unless you explicitly set `method='DM'`, when it takes only a few seconds. I guess the MatrixElement objects create an EX poly? That should perhaps be fixed, but even so, it seems like even for EX domains this method should be preferred. 

Of course, if I instead do

```
Matrix(4, 4, lambda i, j: symbols(f"x_{i}_{j}")).inv()
```

it takes under a second, and I can compute up to 6x6 in a reasonable amount of time. 

We should add some of these cases (at least up to n=4) to the benchmarks.

Related: https://github.com/sympy/sympy/issues/11082