@oscarbenjamin 2023-07-31
> I guess the MatrixElement objects create an EX poly? That should perhaps be fixed, but even so, it seems like even for EX domains this method should be preferred.

Currently `DomainMatrix` uses `construct_domain` which rejects a mixture of `MatrixElement` from the same matrix:
```python
In [11]: construct_domain(symbols('x:4'))
Out[11]: (ZZ[x0,x1,x2,x3], [x0, x1, x2, x3])

In [12]: construct_domain(list(MatrixSymbol('x', 2, 2)))
Out[12]: (EX, [EX(x[0, 0]), EX(x[0, 1]), EX(x[1, 0]), EX(x[1, 1])])
```
They are rejected conervatively because they share free symbols:
```python
In [14]: m1,m2,m3,m4 = MatrixSymbol('x', 2, 2)

In [15]: m1.free_symbols
Out[15]: {x}
```
https://github.com/sympy/sympy/blob/e8ffcf40a6148235667ed75b811bdc1f953466e9/sympy/polys/constructor.py#L151-L157

I have been contemplating switching to use `Poly` rather than `construct_domain` because `Poly` is less strict about generators:
```python
In [16]: Poly(m1+m2+m3+m4)
Out[16]: Poly(x[0, 0] + x[0, 1] + x[1, 0] + x[1, 1], x[0, 0], x[0, 1], x[1, 0], x[1, 1], domain='ZZ')
```
See the comment about `parallel_poly_from_expr` in gh-25448. See also gh-22237