@oscarbenjamin 2023-07-31
> That should perhaps be fixed, but even so, it seems like even for EX domains this method should be preferred.

Maybe. Some of the code for existing matrix functions uses things like `simplify` which are impossible for `DomainMatrix` to replicate in full. That does not matter for a proper domain because the domain already has automatic simplification and canonicalisation. It it hard to know without following through all of the logic whether or not someone is using something like that.

In general it would be better to focus on making the domain system do what we want it to do rather than trying to get DomainMatrix to work with EX. The existing matrix routines basically try to work with EX and perhaps if we take away the cases that can be handled more efficiently by DomainMatrix then those routines can be better optimised to handle the remaining cases.