@oscarbenjamin 2023-10-08
There are still further algorithmic improvements that can be made here but I would like to get some version of this merged because it brings a big speed up for many common cases.

This `rand` function makes an N x N matrix with n different symbols:
```python
In [5]: def rand(N, n): return sum([randMatrix(N)*x for x in symbols(f'x:{n}')], randMatrix(N))

In [6]: rand(2, 3)
Out[6]: 
⎡27⋅x₀ + 35⋅x₁ + 58⋅x₂ + 93  27⋅x₀ + 86⋅x₁ + 31⋅x₂ + 49⎤
⎢                                                      ⎥
⎣10⋅x₀ + 53⋅x₁ + 2⋅x₂ + 80   70⋅x₀ + 81⋅x₁ + 30⋅x₂ + 80⎦

In [7]: rand(5, 1)
Out[7]: 
⎡23⋅x₀ + 20  77⋅x₀ + 5   78⋅x₀ + 30  77⋅x₀ + 62  36⋅x₀ + 16⎤
⎢                                                          ⎥
⎢14⋅x₀ + 15  20⋅x₀ + 38  3⋅x₀ + 53   46⋅x₀ + 1   40⋅x₀ + 51⎥
⎢                                                          ⎥
⎢26⋅x₀ + 21  45⋅x₀ + 45  63⋅x₀ + 19  88⋅x₀ + 52  12⋅x₀ + 97⎥
⎢                                                          ⎥
⎢35⋅x₀ + 95  16⋅x₀ + 36  37⋅x₀ + 57  56⋅x₀ + 72  6⋅x₀ + 76 ⎥
⎢                                                          ⎥
⎣60⋅x₀ + 12  78⋅x₀ + 62  94⋅x₀ + 81  36⋅x₀ + 17  11⋅x₀ + 16⎦
```
Timings with the PR are:
```python
In [1]: from time import time

In [8]: for n in range(1, 10):
   ...:     M = rand(n, 1)
   ...:     start = time()
   ...:     Minv = M.inv()
   ...:     time_taken = time() - start
   ...:     print(f'{n}x{n} matrix 1 symbol: {time_taken:.3g} seconds')
   ...: 
1x1 matrix 1 symbol: 0.00432 seconds
2x2 matrix 1 symbol: 0.0101 seconds
3x3 matrix 1 symbol: 0.0299 seconds
4x4 matrix 1 symbol: 0.0788 seconds
5x5 matrix 1 symbol: 0.116 seconds
6x6 matrix 1 symbol: 0.167 seconds
7x7 matrix 1 symbol: 0.271 seconds
8x8 matrix 1 symbol: 0.317 seconds
9x9 matrix 1 symbol: 0.44 seconds
```
With master the equivalent timings are:
```python
1x1 matrix 1 symbol: 0.377 seconds
2x2 matrix 1 symbol: 0.328 seconds
3x3 matrix 1 symbol: 1.15 seconds
4x4 matrix 1 symbol: 5.67 seconds
5x5 matrix 1 symbol: 25.6 seconds
^C^C
KeyboardInterrupt
```
With 2 symbols we have:
```python
# PR
1x1 matrix 2 symbol: 0.00853 seconds
2x2 matrix 2 symbol: 0.0176 seconds
3x3 matrix 2 symbol: 0.0605 seconds
4x4 matrix 2 symbol: 0.161 seconds
5x5 matrix 2 symbol: 0.24 seconds
6x6 matrix 2 symbol: 0.441 seconds
7x7 matrix 2 symbol: 0.892 seconds
8x8 matrix 2 symbol: 1.73 seconds
9x9 matrix 2 symbol: 2.9 seconds

# Master
1x1 matrix 2 symbol: 0.0829 seconds
2x2 matrix 2 symbol: 0.635 seconds
3x3 matrix 2 symbol: 7.28 seconds
^C^C
KeyboardInterrupt
```
Plenty more timings can be done but I think that just being able to invert a simple 3x3 matrix with some symbols in it is a common enough case.

This also fixes gh-19920.